% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doRedis-package.R
\docType{package}
\name{doRedis-package}
\alias{doRedis-package}
\title{A Redis parallel back end for foreach.}
\description{
The doRedis package imlpements an elastic parallel back end
for foreach using the Redis key/value database.
}

