% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doRedis.R
\name{.doRedis}
\alias{.doRedis}
\title{internal function called by foreach}
\usage{
.doRedis(obj, expr, envir, data)
}
\arguments{
\item{obj}{a foreach object}

\item{expr}{the expression to evaluate}

\item{envir}{the expression environment}

\item{data}{a list of parameters from registerDoRedis}
}
\value{
the foreach result
}
\description{
internal function called by foreach
}
