% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doRedis.R
\name{jobs}
\alias{jobs}
\title{List doRedis jobs}
\usage{
jobs(queue = "*")
}
\arguments{
\item{queue}{List jobs for the specified queue, or set to "*" to list jobs for all queues}
}
\value{
a data frame listing jobs by row with variables queue, id, user, host and time (submitted).
}
\description{
List doRedis jobs
}
