% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doRedis.R
\name{logger}
\alias{logger}
\title{Print a timestamped message to the  standard error stream.}
\usage{
logger(msg)
}
\arguments{
\item{msg}{a character message to print to the standard error stream}
}
\value{
The character string that was printed, decorated with time and system info.
}
\description{
Use to help debug remote doRedis workers.
}
