% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redux.R
\name{redisConnect}
\alias{redisConnect}
\title{Internal redux compatability functions with old rredis package}
\usage{
redisConnect(host = "localhost", port = 6379, password, ...)
}
\description{
Internal redux compatability functions with old rredis package
}

