% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redux.R
\name{redisDelete}
\alias{redisDelete}
\title{A convenience function to delete a Redis key}
\usage{
redisDelete(key)
}
\arguments{
\item{key}{(character or raw) Redis key name to delete}
}
\value{
Redis status message
}
\description{
A convenience function to delete a Redis key
}
\seealso{
\code{\link{hiredis}}
}
