% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redux.R
\name{redisSet}
\alias{redisSet}
\title{A convenience function to set an R value in a Redis key}
\usage{
redisSet(key, val)
}
\arguments{
\item{key}{(character or raw) Redis key name}

\item{val}{R value to set}
}
\value{
Redis status message
}
\description{
This function serializes the val argument.
}
\seealso{
\code{\link{hiredis}}
}
