% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redisWorker.R
\name{redisWorker}
\alias{redisWorker}
\title{Initialize a doRedis worker process.}
\usage{
redisWorker(queue, host = "localhost", port = 6379, iter = Inf,
  linger = 30, log = stderr(), connected = FALSE, password = NULL,
  loglevel = 0, ...)
}
\arguments{
\item{queue}{work queue name or a vector of queue names}

\item{host}{Redis database host name or IP address}

\item{port}{Redis database port number}

\item{iter}{maximum number of tasks to process before exiting the worker loop}

\item{linger}{timeout in seconds after which the work queue is deleted that the worker terminates}

\item{log}{print messages to the specified file connection}

\item{connected}{set to \code{TRUE} to reuse an existing open connection to Redis, otherwise establish a new one}

\item{password}{optional Redis database password}

\item{loglevel}{set to > 0 to increase verbosity in the log}

\item{...}{Optional additional parameters passed to \code{\link{redisConnect}}}
}
\value{
NULL is invisibly returned.
}
\description{
The redisWorker function enrolls the current R session in one or
more doRedis worker pools specified by the work queue names. The worker
loop takes over the R session until the work queue(s) are deleted, after
which at most \code{linger} seconds the worker loop exits, or until
the worker has processed \code{iter} tasks.
}
\note{
The worker connection to Redis uses a TCP timeout value of 30 seconds by
default. That means that the worker will exit after about 30 seconds of inactivity.
If you want the worker to remain active for longer periods, use the \code{timeout}
option, for instance setting \code{timeout=0} tells the worker to listen indefinitely
for work.

Use the \code{linger} option to instruct the worker to linger for up to the indicated
number of seconds after the listening work queue has been removed. After at most that
interval, the worker will exit after removing the queue.
}
\seealso{
\code{\link{registerDoRedis}}, \code{\link{startLocalWorkers}}
}

