% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doRedis.R
\name{removeQueue}
\alias{removeQueue}
\title{Remove a doRedis queue and delete all associated keys from Redis.}
\usage{
removeQueue(queue)
}
\arguments{
\item{queue}{The doRedis queue name}
}
\value{
NULL is invisibly returned.
}
\description{
Removing a doRedis queue cleans up associated keys in the Redis
database and signals to workers listening on the queue to terminate.
Workers normally terminate after their timeout period after a
queue is delete.
}
\note{
Workers listening for work on more than one queue will only
terminate after all their queues have been deleted.
}

