% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doRedis.R
\name{setGather}
\alias{setGather}
\title{Set distributed accumulation}
\usage{
setGather(value = FALSE)
}
\arguments{
\item{value}{logical value; the default value is \code{FALSE}.}
}
\value{
\code{NULL}
}
\description{
When \code{TRUE}, run the \code{.combine} function on workers prior to gathering
results at the master. Only affects runs with \code{chunkSize} greater
than one.
}
\note{
This value is overriden by setting the 'gather' option in the
foreach loop (see the examples).

Be aware that some \code{.combine} functions may perform differently when
their invocation is distributed.
}
\examples{
\dontrun{
setChunkSize(5)
setGather(TRUE)
foreach(j=1:10, .combine=c) \%dopar\% j

# Same effect as:

foreach(j=1:10, .combine=c,
        .options.redis=list(chunksize=5, gather=TRUE)) \%dopar\% j
}

}

