% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doRedis.R
\name{setProgress}
\alias{setProgress}
\title{Progress bar}
\usage{
setProgress(value = FALSE)
}
\arguments{
\item{value}{if \code{TRUE}, display a text progress bar indicating status of the computation}
}
\value{
\code{value} is invisibly returned (this function is used for its side effect).
}
\description{
Progress bar
}
