% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doRedis.R
\name{setStream}
\alias{setStream}
\title{Set experimental streaming task submission}
\usage{
setStream(value = FALSE)
}
\arguments{
\item{value}{\code{TRUE} or \code{FALSE}}
}
\value{
\code{value} is invisibly returned.
}
\description{
The \code{setStream} function overrides the same option specified
inline in the \code{foreach} statement.
}
\examples{
\dontrun{
setStream(TRUE)
foreach(j=1:10) \%dopar\% j

# Same effect as:

foreach(j=1:10,
        .options.redis=list(stream=TRUE)) \%dopar\% j
}
}

