% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redux.R
\name{redisConnect}
\alias{redisConnect}
\title{Explicitly connect to a Redis server.}
\usage{
redisConnect(host = "localhost", port = 6379L, password, ...)
}
\arguments{
\item{host}{character Redis host name}

\item{port}{integer Redis port number}

\item{password}{optional character Redis password}

\item{...}{optional additional arguments for compatability with old rredis, ignored}
}
\description{
This function is normally not needed, use the redux package functions instead,
or simply registerDoRedis.
}
\seealso{
\code{\link{registerDoRedis}}, \code{\link{redisWorker}}, \code{\link{startLocalWorkers}}
}
