% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redisWorker.R
\name{redisWorker}
\alias{redisWorker}
\title{Initialize a doRedis worker process.}
\usage{
redisWorker(queue, host = "localhost", port = 6379, iter = Inf,
  timeout = 30, log = stderr(), connected = FALSE, password = NULL,
  loglevel = 0, ...)
}
\arguments{
\item{queue}{work queue name or a vector of queue names}

\item{host}{Redis database host name or IP address}

\item{port}{Redis database port number}

\item{iter}{maximum number of tasks to process before exiting the worker loop}

\item{timeout}{timeout in seconds after which the work queue is deleted that the worker terminates}

\item{log}{print messages to the specified file connection}

\item{connected}{Is the R session creating the worker already connected to Redis?}

\item{password}{optional Redis database password}

\item{loglevel}{set to 1 to increase verbosity in the log}

\item{...}{Optional additional parameters passed to \code{\link{redisConnect}}}
}
\value{
NULL is invisibly returned.
}
\description{
The redisWorker function enrolls the current R session in one or
more doRedis worker pools specified by the work queue names. The worker
loop takes over the R session until the work queue(s) are deleted, after
which at most \code{timeout} seconds the worker loop exits, or until
the worker has processed \code{iter} tasks.
}
\seealso{
\code{\link{registerDoRedis}}, \code{\link{startLocalWorkers}}
}

