% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/redisWorker.R
\name{redisWorker}
\alias{redisWorker}
\title{Initialize a doRedis worker process.}
\usage{
redisWorker(queue, host = "localhost", port = 6379, iter = Inf,
  timeout = 30, log = stdout(), connected = FALSE, password = NULL, ...)
}
\arguments{
\item{queue}{The doRedis work queue name or a vector of queue names.}

\item{host}{The Redis database host name or IP address.}

\item{port}{The Redis database port number.}

\item{iter}{Maximum number of tasks to process before exiting the worker loop.}

\item{timeout}{Timeout in seconds after which the work queue is deleted that the worker terminates.}

\item{log}{Log messages to the specified file connection.}

\item{connected}{Is the R session creating the worker already connected to Redis?}

\item{password}{Optional Redis database password.}

\item{...}{Optional additional options to pass along to redisConnect.}
}
\value{
NULL is invisibly returned.
}
\description{
The redisWorker function enrolls the current R session in one or
more doRedis worker pools specified by the work queue names. The worker
loop takes over the R session until the work queue(s) are deleted, after
which at most \code{timeout} seconds the worker loop exits, or until
the worker has processed \code{iter} tasks.
}
\seealso{
\code{\link{registerDoRedis}}, \code{\link{startLocalWorkers}}
}

