% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doRedis.R
\name{removeJob}
\alias{removeJob}
\title{Remove Redis keys associated with one or more doRedis jobs}
\usage{
removeJob(job)
}
\arguments{
\item{job}{Either a named character vector with "queue" and "id" entries corresponding to a doRedis
job queue and job id, or a list with equal-length "queue" and "id" entries, or a data frame with
"queue" and "id" entries, for example as returned by \code{\link{jobs}}.}
}
\value{
NULL is invisibly returned
}
\description{
Remove Redis keys associated with one or more doRedis jobs
}
