% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doRedis.R
\name{setGather}
\alias{setGather}
\title{Set distributed reduction}
\usage{
setGather(fun = NULL)
}
\arguments{
\item{fun}{a function of two arguments, set to NULL to disable gather}
}
\value{
\code{NULL}
}
\description{
Instruct doRedis to perform the \code{.combine} reduction per task on the
workers before returning results. Combined results are then processed through
the specified \code{fun} function. The gather function \code{fun} is an
'outer' combine function.  Only applies when \code{chunkSize} greater than
one, and implies that \code{.multicombine=FALSE}.
}
\note{
This value is overriden by setting the 'gather' option in the
foreach loop (see the examples).
}
\examples{
\dontrun{
setChunkSize(3)
setGather(list)
foreach(j=1:10, .combine=c) \%dopar\% j

# Same effect as:

foreach(j=1:10, .combine=c,
        .options.redis=list(chunksize=3, gather=list)) \%dopar\% j
}
}
\seealso{
\code{\link{foreach}}, \code{\link{setChunkSize}}
}

