% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doRedis.R
\name{setPackages}
\alias{setPackages}
\title{Manually set package names in the worker environment package list.}
\usage{
setPackages(packages = c())
}
\arguments{
\item{packages}{A character vector of package names.}
}
\value{
The value of \code{packages} is invisibly returned.
}
\description{
The \code{setPackages} function lets users manually declare packages
that R worker processes need to load before running their tasks.
}
\details{
The \code{foreach} function includes a similar \code{.packages} parameter.

Defines a way to set the foreach \code{.packages} option for users without direct access
to the \code{foreach} function, for example, when \code{foreach} is used
inside another package.
}
